+++ CHANGELOG.md
@@ -6,7 +6,10 @@ All notable changes to SmartSpend will be documented in this file.
 
 
 
+
 ## [unreleased] - {relase date}
+
+## [0.16.13] - 2026-02-26
 
 - **fix(ui): increase home page bottom padding so FAB doesn't overlap last transaction**